test(core): split agent profile SDK test responsibilities
(cherry picked from commit 84bc92f12cebd1978ae4b1d670e47bd5843cdc28)
